chore: disable tests that fail because of broken ts-config
