+++ frontend/src/pages/webcomponent/_id/_version/index.vue
